+++ 03_Implementierung/GO-Tomcat/src/main/java/edu/kit/pse17/go_app/ServiceLayer/UserLocation.java
@@ -2,14 +2,14 @@ package edu.kit.pse17.go_app.ServiceLayer;
 
 
 /**
- * Bei dieser Klasse handelt es sich um eine Datenhaltungsklasse, die dem Clustering-Algorithmus das hantieren mit den Standorten erleichtert.
- * Ein Objekt dieser Klasse beschreibt dabei einen Benutzerstandort.
+ * Bei dieser Klasse handelt es sich um eine Datenhaltungsklasse, die dem Clustering-Algorithmus das hantieren mit den
+ * Standorten erleichtert. Ein Objekt dieser Klasse beschreibt dabei einen Benutzerstandort.
  */
 public class UserLocation {
 
     /**
-     * Die ID des Benutzers, um dessen Standort es sich handelt. Es handelt sich um eine gültige Benutzer-ID, die von anderen Klassen der anwendung erkannt
-     * werden kann.
+     * Die ID des Benutzers, um dessen Standort es sich handelt. Es handelt sich um eine gültige Benutzer-ID, die von
+     * anderen Klassen der anwendung erkannt werden kann.
      */
     private String userId;
 
@@ -17,24 +17,25 @@ public class UserLocation {
      * Der geographische Breitengrad des Standorts des Benutzers. Der Wert muss als Breitengrad interpretierbar sein,
      * muss also zwischen +90 und -90 liegen.
      */
-    private long lat;
+    private double lat;
 
     /**
      * Der geographische Längengrad des Standorts des Benutzers. Der Wert muss als Längengrad interpretierbar sein,
      * muss also zwischen +180 und -180 liegen.
      */
-    private long lon;
+    private double lon;
 
     /**
      * Ein Konstruktor, der sämtliche Attribute der Klasse als Argumente entgegennimmt und daraus ein Objekt erzeugt.
-     * @param userId Die ID des Benutzers, um dessen Standort es sich handelt. Es handelt sich um eine gültige Benutzer-ID, die von anderen Klassen der anwendung erkannt
-     * werden kann.
-     * @param lat Der geographische Breitengrad des Standorts des Benutzers. Der Wert muss als Breitengrad interpretierbar sein,
-     * muss also zwischen +90 und -90 liegen.
-     * @param lon Der geographische Längengrad des Standorts des Benutzers. Der Wert muss als Längengrad interpretierbar sein,
-     * muss also zwischen +180 und -180 liegen.
+     *
+     * @param userId Die ID des Benutzers, um dessen Standort es sich handelt. Es handelt sich um eine gültige
+     *               Benutzer-ID, die von anderen Klassen der anwendung erkannt werden kann.
+     * @param lat    Der geographische Breitengrad des Standorts des Benutzers. Der Wert muss als Breitengrad
+     *               interpretierbar sein, muss also zwischen +90 und -90 liegen.
+     * @param lon    Der geographische Längengrad des Standorts des Benutzers. Der Wert muss als Längengrad
+     *               interpretierbar sein, muss also zwischen +180 und -180 liegen.
      */
-    public UserLocation(String userId, long lat, long lon) {
+    public UserLocation(String userId, double lat, double lon) {
         this.userId = userId;
         this.lat = lat;
         this.lon = lon;
@@ -48,19 +49,19 @@ public class UserLocation {
         this.userId = userId;
     }
 
-    public long getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(long lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public long getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(long lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
 }
